+++ R/make_gloss_list.R
@@ -7,6 +7,7 @@
 #' @param all_possible_variants logical. Some glosses have multiple definitions.
 #' @return a string with glosses and their definitions gathered from \code{definition_source} table.
 #' @importFrom knitr asis_output
+#' @importFrom knitr opts_current
 #' @importFrom rmarkdown metadata
 #' @importFrom utils read.csv
 #' @export
@@ -14,6 +15,9 @@
 make_gloss_list <- function(definition_source = lingglosses::glosses_df,
                             all_possible_variants = FALSE){
 
+  knitr::opts_current$set(results='asis')
+
+# checks for the user's dataset --------------------------------------------
   if(!("data.frame" %in% attr(definition_source, "class"))){
     stop("Argument 'definition_source' should have an attribute 'data.frame'.")
   }
@@ -30,27 +34,38 @@ make_gloss_list <- function(definition_source = lingglosses::glosses_df,
     definition_source$source <- ""
   }
 
+# get glosses --------------------------------------------------------------
 
-  knitr::opts_current$set(results='asis')
-
-  # get glosses
   gloss_list <- unlist(utils::read.csv(getOption("lingglosses.glosses_list")))
-  # get definitions and sort them
-  gloss_ld <- definition_source[definition_source$gloss %in% gloss_list,]
-  if(length(gloss_list[!(gloss_list %in% definition_source$gloss)]) > 0){
-    gloss_ld <- unique(rbind(
-      gloss_ld,
-      data.frame(gloss = gloss_list[!(gloss_list %in% definition_source$gloss)],
+
+# create a temporal file with all glosses that will be modified ------------
+  glosses_dataset <- lingglosses::glosses_df[lingglosses::glosses_df$gloss %in%
+                                               gloss_list, ]
+
+# change definition from lingglosses::glosses to user's values -------------
+  res <- lapply(seq_along(definition_source$gloss), function(x){
+    glosses_dataset[glosses_dataset$gloss ==
+                             definition_source$gloss[x],
+                           "definition"] <<- definition_source$definition[x]
+  })
+
+# keep only those that are present in the document -------------------------
+  glosses_dataset <- glosses_dataset[glosses_dataset$gloss %in% gloss_list, ]
+
+# for those glosses that are not present in our and user's database --------
+  if(length(gloss_list[!(gloss_list %in% glosses_dataset$gloss)]) > 0){
+    glosses_dataset <- unique(rbind(
+      glosses_dataset,
+      data.frame(gloss = gloss_list[!(gloss_list %in% glosses_dataset$gloss)],
                  definition = "",
                  source = "",
                  weight = 1)))
   }
 
-  # it is sorted in lingglosses::glosses, but may be not sorted in user's tables
-  gloss_ld <- gloss_ld[order(gloss_ld$gloss),]
+  glosses_dataset <- glosses_dataset[order(glosses_dataset$gloss),]
 
   if(isFALSE(all_possible_variants)){
-    gloss_ld <- gloss_ld[gloss_ld$weight != 0,]
+    glosses_dataset <- glosses_dataset[glosses_dataset$weight != 0,]
   }
 
   # generate non breacking space
@@ -62,8 +77,8 @@ make_gloss_list <- function(definition_source = lingglosses::glosses_df,
   }
 
   # create an output
-  res <- paste(lingglosses::small_caps(gloss_ld$gloss),
-               gloss_ld$definition,
+  res <- paste(lingglosses::small_caps(glosses_dataset$gloss),
+               glosses_dataset$definition,
                sep = gloss_sep, collapse = "; ")
   knitr::asis_output(res)
 }